backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
pydantic[email]
email-validator
bcrypt
python-multipart